refactor: migrate my_blocks double-conversion tests to unit roundtrip tests
Add double-conversion tests to ruby-roundtrip/my_blocks.test.js by
calling rubyToBlocksToRuby twice sequentially — matching the
Ruby→Blocks→Ruby→Blocks→Ruby flow that the integration test verified
via browser (Selenium).

Remove the last remaining integration test file in ruby-tab/:
  - test/integration/ruby-tab/my_blocks.test.js

All ruby-tab integration tests are now fully replaced by unit tests.
