+++ packages/scratch-gui/test/unit/lib/ruby-roundtrip/my_blocks.test.js
@@ -1,13 +1,13 @@
 /**
  * Unit test replacing test/integration/ruby-tab/my_blocks.test.js
- * (roundtrip-compatible tests only; double-conversion UI tests are omitted)
  */
 import dedent from 'dedent';
 import {
     makeSpriteTarget,
     makeConverter,
     setupRubyGenerator,
-    expectRoundTrip
+    expectRoundTrip,
+    rubyToBlocksToRuby
 } from '../../helpers/ruby-roundtrip-helper';
 
 describe('Ruby Roundtrip: My Blocks (procedure) category', () => {
@@ -79,5 +79,43 @@ describe('Ruby Roundtrip: My Blocks (procedure) category', () => {
               say(@_return_add)
             end
         `);
+    });
+
+    test('Double conversion should not lose method body: minimal', async () => {
+        const code = dedent`
+            def self.add
+              move(10)
+            end
+        `;
+        // First conversion: Ruby -> Blocks -> Ruby
+        const firstResult = await rubyToBlocksToRuby(converter, target, code);
+        expect(firstResult).toContain('def self.add');
+        expect(firstResult).toContain('move(10)');
+        expect(firstResult).not.toMatch(/def self\.add\s*end/);
+
+        // Second conversion: Ruby -> Blocks -> Ruby again (THIS IS WHERE THE BUG OCCURS)
+        const secondResult = await rubyToBlocksToRuby(converter, target, firstResult);
+        expect(secondResult).toContain('def self.add');
+        expect(secondResult).toContain('move(10)');
+        expect(secondResult).not.toMatch(/def self\.add\s*end/);
+    });
+
+    test('Double conversion should not lose method body: return value', async () => {
+        const code = dedent`
+            def self.add(a, b)
+              a + b
+            end
+        `;
+        // First conversion: Ruby -> Blocks -> Ruby
+        const firstResult = await rubyToBlocksToRuby(converter, target, code);
+        expect(firstResult).toContain('def self.add');
+        expect(firstResult).toContain('a + b');
+        expect(firstResult).not.toMatch(/def self\.add\([^)]*\)\s*end/);
+
+        // Second conversion: Ruby -> Blocks -> Ruby again (THIS IS WHERE THE BUG OCCURS)
+        const secondResult = await rubyToBlocksToRuby(converter, target, firstResult);
+        expect(secondResult).toContain('def self.add');
+        expect(secondResult).toContain('a + b');
+        expect(secondResult).not.toMatch(/def self\.add\([^)]*\)\s*end/);
     });
 });
